Extract ring-buffer into standalone crate
