+++ Cargo.toml
@@ -52,6 +52,7 @@ members = [
     "kernel/libs/jhash",
     "kernel/libs/keyable-arc",
     "kernel/libs/logo-ascii-art",
+    "kernel/libs/ring-buffer",
     "kernel/libs/typeflags",
     "kernel/libs/typeflags-util",
     "kernel/libs/xarray",
@@ -88,6 +89,7 @@ default-members = [
     "kernel/libs/aster-bigtcp",
     "kernel/libs/aster-util",
     "kernel/libs/device-id",
+    "kernel/libs/ring-buffer",
     "kernel/libs/xarray",
 ]
 # All user-space or host-native crates.
@@ -164,6 +166,7 @@ device-id = { path = "kernel/libs/device-id" }
 jhash = { path = "kernel/libs/jhash" }
 keyable-arc = { path = "kernel/libs/keyable-arc" }
 logo-ascii-art = { path = "kernel/libs/logo-ascii-art" }
+ring-buffer = { path = "kernel/libs/ring-buffer" }
 typeflags = { path = "kernel/libs/typeflags" }
 typeflags-util = { path = "kernel/libs/typeflags-util" }
 xarray = { path = "kernel/libs/xarray" }
