+++ kernel/Cargo.toml
@@ -56,6 +56,7 @@ rand = { workspace = true, features = [
     "small_rng",
     "std_rng",
 ] }
+ring-buffer.workspace = true
 spin.workspace = true
 takeable.workspace = true
 time.workspace = true
